+++ building_sim_plugins/building_plugins_common/src/crowd_simulator_common.cpp
@@ -220,7 +220,8 @@ void CrowdSimInterface::_add_object(AgentPtr agent_ptr,
     assert(!model_name.empty());
   }
   _objects.emplace_back(
-    new Object{agent_ptr, model_name, type_name, is_external});
+    new Object{agent_ptr, model_name, type_name, is_external,
+      AnimState::WALK});
 }
 
 size_t CrowdSimInterface::get_num_objects() const
@@ -238,6 +239,28 @@ const
 void CrowdSimInterface::one_step_sim() const
 {
   _menge_handle->sim_step();
+}
+
+double CrowdSimInterface::get_switch_anim_distance_th() const
+{
+  return _switch_anim_distance_th;
+}
+
+std::vector<std::string> CrowdSimInterface::get_switch_anim_name() const
+{
+  return _switch_anim_name;
+}
+
+//=============================================
+CrowdSimInterface::AnimState CrowdSimInterface::Object::get_next_state(
+  bool condition)
+{
+  if (condition)
+    return AnimState::IDLE;
+  else
+    return AnimState::WALK;
+
+  return current_state;
 }
 
 } //namespace crowd_simulator