+++ traffic_editor/gui/project.h
@@ -120,14 +120,15 @@ public:
     const int start_idx,
     const int end_idx);
 
+#ifdef HAS_IGNITION_PLUGIN
   // simulation stuff
   void sim_reset();
   void sim_tick();
   bool sim_is_paused = true;
+  bool has_sim_plugin();
+#endif
 
   RenderingOptions rendering_options;
-
-  bool has_sim_plugin();
 
   bool set_filename(const std::string& _filename);
   std::string get_filename() { return filename; }